+++ zebra-consensus/src/block.rs
@@ -8,7 +8,6 @@
 //! verification, where it may be accepted or rejected.
 
 use std::{
-    collections::HashSet,
     future::Future,
     pin::Pin,
     sync::Arc,
@@ -26,7 +25,7 @@ use zebra_chain::{
     amount::Amount,
     block,
     parameters::{subsidy::SubsidyError, Network},
-    transaction, transparent,
+    transparent,
     work::equihash,
 };
 use zebra_state as zs;
@@ -291,9 +290,6 @@ where
                 &transaction_hashes,
             ));
 
-            let known_outpoint_hashes: Arc<HashSet<transaction::Hash>> =
-                Arc::new(known_utxos.keys().map(|outpoint| outpoint.hash).collect());
-
             for (&transaction_hash, transaction) in
                 transaction_hashes.iter().zip(block.transactions.iter())
             {
@@ -304,7 +300,6 @@ where
                     .call(tx::BlockRequest {
                         transaction_hash,
                         transaction: transaction.clone(),
-                        known_outpoint_hashes: known_outpoint_hashes.clone(),
                         known_utxos: known_utxos.clone(),
                         height,
                         time: block.header.time,